refactor: move directory helper to logger
- Move `ensureDirectoryExists` function from `config.ts` to `logger.ts`.
- This utility is only used for creating log file directories.
- Call the helper during logger initialization to ensure log paths exist before use.
